Move DexParser to util package
