+++ core/src/main/java/org/lsposed/lspd/impl/LSPosedContext.java
@@ -66,6 +66,7 @@ import de.robv.android.xposed.XposedInit;
 import io.github.libxposed.XposedContext;
 import io.github.libxposed.XposedModule;
 import io.github.libxposed.XposedModuleInterface;
+import io.github.libxposed.util.DexParser;
 
 public class LSPosedContext extends XposedContext {
