+++ libxposed/api/src/main/java/io/github/libxposed/XposedContext.java
@@ -5,4 +5,3 @@ import android.content.Context;
 public abstract class XposedContext extends Context implements XposedInterface {
 
 }
-
