+++ libxposed/api/src/main/java/io/github/libxposed/XposedContextWrapper.java
@@ -11,6 +11,8 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 import java.nio.ByteBuffer;
 
+import io.github.libxposed.util.DexParser;
+
 public class XposedContextWrapper extends ContextWrapper implements XposedInterface {
 
     XposedContextWrapper(XposedContext base) {
